+++ config/oembed.php
@@ -1,0 +1,10 @@
+<?php
+
+/**
+ * Configuration file for oEmbed
+ * https://github.com/wrav/oembed
+ */
+
+return [
+    'enableCache' => true,
+];